Convert controller to TypeScript
